리팩토링: RestController com.gunyoung.info.controller.rest 패키지로 분리(ContentRestController, PersonRestController, ProfileRestController)
