+++ src/main/java/com/gunyoung/info/controller/rest/RestfulController.java
@@ -1,10 +1,10 @@
-package com.gunyoung.info.controller;
+package com.gunyoung.info.controller.rest;
 
 import java.util.LinkedList;
 import java.util.List;
 
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RestController;
 
 import com.gunyoung.info.domain.Person;
@@ -32,8 +32,7 @@ public class RestfulController {
 	 *  </pre>
 	 *  @author kimgun-yeong
 	 */
-	
-	@GetMapping("/main/list")
+	@RequestMapping(value="/main/list",method=RequestMethod.GET)
 	public List<MainListObject> index() {
 		List<Person> personList = personService.getAll();
 		
@@ -44,19 +43,5 @@ public class RestfulController {
 		}
 		
 		return result;
-	}
-	
-	/**
-	 * <pre>
-	 *  - 기능: 회원가입할때 Email 중복 여부 반환하는 컨트롤러
-	 *  - 반환: 
-	 *  	True or False
-	 *  </pre>
-	 *  @param email 중복여부를 확인하려는 email 값
-	 *  @author kimgun-yeong
-	 */
-	@GetMapping("/join/emailverification")
-	public String emailVerification(@RequestParam("email") String email) {
-		return String.valueOf(personService.existsByEmail(email));
 	}
 }
